model-inference/decisionTree/experiments/requirements.txt
absl-py==1.0.0
anyio==3.5.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cmdstanpy==0.9.68
colorama==0.4.3
command-not-found==0.3
connectorx==0.2.5
convertdate==2.4.0
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
ephem==4.1.3
et-xmlfile==1.1.0
executing==0.8.3
fasteners==0.14.1
filelock==3.6.0
flatbuffers==1.12
fonttools==4.30.0
future==0.18.2
gast==0.5.3
gdown==4.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
hijri-converter==2.2.3
holidays==0.13
httplib2==0.14.0
Hummingbird==0.0.3
hummingbird-ml==0.4.4
idna==2.8
importlib-metadata==4.11.2
importlib-resources==5.4.0
inform==1.26.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.0
korean-lunar-calendar==0.2.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
lightgbm==3.3.2
lockfile==0.12.2
louis==3.12.0
LunarCalendar==0.0.9
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
memory-profiler==0.60.0
mistune==0.8.4
monotonic==1.5
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
nose==1.3.7
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
onnx==1.11.0
onnxconverter-common==1.9.0
onnxmltools==1.11.0
onnxruntime-gpu==1.11.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickle-compat==2.1.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.5.1
plotly==5.8.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
prophet==1.0.1
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMeeus==0.5.11
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.2.1
pystan==2.19.1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==6.1.2
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
quantiphy==2.17.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
setuptools-git==1.2
simplejson==3.16.0
six==1.14.0
skl2onnx==1.11.1
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
systemd-python==234
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-decision-forests==0.2.4
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf2onnx==1.9.3
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.63.2
traitlets==5.1.1
typing-extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.2.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.14.0
wurlitzer==3.0.2
xgboost==1.6.0
xkit==0.0.0
zipp==3.7.0
